Move hash related utils, brave wallet types, Uint256<=>Hex to common
