+++ components/brave_wallet/browser/eth_data_builder.h
@@ -9,7 +9,7 @@
 #include <string>
 #include <vector>
 #include "base/values.h"
-#include "brave/components/brave_wallet/browser/brave_wallet_types.h"
+#include "brave/components/brave_wallet/common/brave_wallet_types.h"
 
 namespace brave_wallet {
