+++ components/brave_wallet/browser/eth_data_builder_unittest.cc
@@ -5,7 +5,7 @@
 
 #include "brave/components/brave_wallet/browser/eth_data_builder.h"
 
-#include "brave/components/brave_wallet/browser/brave_wallet_utils.h"
+#include "brave/components/brave_wallet/common/hex_utils.h"
 #include "testing/gtest/include/gtest/gtest.h"
 
 namespace brave_wallet {
