+++ components/brave_wallet/browser/eth_response_parser.cc
@@ -12,6 +12,7 @@
 #include "base/strings/string_number_conversions.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_utils.h"
 #include "brave/components/brave_wallet/browser/eth_address.h"
+#include "brave/components/brave_wallet/common/hex_utils.h"
 
 namespace {
