+++ components/brave_wallet/browser/eth_response_parser.h
@@ -10,7 +10,7 @@
 #include <vector>
 
 #include "base/values.h"
-#include "brave/components/brave_wallet/browser/brave_wallet_types.h"
+#include "brave/components/brave_wallet/common/brave_wallet_types.h"
 
 namespace brave_wallet {
