+++ components/brave_wallet/browser/test/BUILD.gn
@@ -13,7 +13,6 @@ source_set("brave_wallet_unit_tests") {
     sources = [
       "//brave/components/brave_wallet/browser/asset_ratio_controller_unittest.cc",
       "//brave/components/brave_wallet/browser/asset_ratio_response_parser_unittest.cc",
-      "//brave/components/brave_wallet/browser/brave_wallet_types_unittest.cc",
       "//brave/components/brave_wallet/browser/brave_wallet_utils_unittest.cc",
       "//brave/components/brave_wallet/browser/eip1559_transaction_unittest.cc",
       "//brave/components/brave_wallet/browser/eip2930_transaction_unittest.cc",
