+++ components/brave_wallet/common/hash_utils.cc
@@ -1,0 +1,46 @@
+/* Copyright (c) 2021 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#include "brave/components/brave_wallet/common/hash_utils.h"
+
+#include <algorithm>
+
+#include "base/strings/string_split.h"
+#include "brave/components/brave_wallet/common/hex_utils.h"
+#include "brave/third_party/ethash/src/include/ethash/keccak.h"
+
+namespace brave_wallet {
+
+std::string KeccakHash(const std::string& input, bool to_hex) {
+  std::vector<uint8_t> bytes(input.begin(), input.end());
+  std::vector<uint8_t> result = KeccakHash(bytes);
+  std::string result_str(result.begin(), result.end());
+  return to_hex ? ToHex(result_str) : result_str;
+}
+
+std::vector<uint8_t> KeccakHash(const std::vector<uint8_t>& input) {
+  auto hash = ethash_keccak256(input.data(), input.size());
+  return std::vector<uint8_t>(hash.bytes, hash.bytes + 32);
+}
+
+std::string GetFunctionHash(const std::string& input) {
+  std::string result = KeccakHash(input);
+  return result.substr(0, std::min(static_cast<size_t>(10), result.length()));
+}
+
+std::string Namehash(const std::string& name) {
+  std::string hash(32, '\0');
+  std::vector<std::string> labels =
+      SplitString(name, ".", base::KEEP_WHITESPACE, base::SPLIT_WANT_NONEMPTY);
+
+  for (auto rit = labels.rbegin(); rit != labels.rend(); rit++) {
+    std::string label_hash = KeccakHash(*rit, false);
+    hash = KeccakHash(hash + label_hash, false);
+  }
+
+  return ToHex(hash);
+}
+
+}  // namespace brave_wallet
